# docs/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Varka Backend API
  version: 0.0.1
  description: REST API for authentication, projects, CIM/PIM/PSM models, transformations, artifacts, modeling metadata, and diagram layout.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1:8080
    description: Local development
security:
  - AuthToken: []
...
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Created auth session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Create an auth session
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Created auth session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [Auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeRequest"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserDto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: End the current auth session
      responses:
        "200":
          description: Session ended
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/artifact:
    get:
      tags: [Artifacts]
      summary: List artifacts for a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artifacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ArtifactRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/artifact/{id}:
    parameters:
      - $ref: "#/components/parameters/ArtifactId"
    get:
      tags: [Artifacts]
      summary: Get an artifact
      responses:
        "200":
          description: Artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/artifact/{id}/file:
    parameters:
      - $ref: "#/components/parameters/ArtifactId"
      - name: path
        in: query
        required: true
        schema:
          type: string
    get:
      tags: [Artifacts]
      summary: Read an artifact file
      responses:
        "200":
          description: File contents
          content:
            text/plain:
              schema:
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/artifact/{id}/files:
    parameters:
      - $ref: "#/components/parameters/ArtifactId"
    put:
      tags: [Artifacts]
      summary: Save or replace an artifact file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveFileRequest"
      responses:
        "200":
          description: Updated artifact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/artifact/{id}/download:
    parameters:
      - $ref: "#/components/parameters/ArtifactId"
    get:
      tags: [Artifacts]
      summary: Download an artifact as a ZIP archive
      responses:
        "200":
          description: ZIP archive
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/impact/artifact/{artifactId}:
    get:
      tags: [Impact]
      summary: Analyze generated artifact lineage
      parameters:
        - name: artifactId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artifact metadata and ordered upstream model ancestors
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    AuthToken:
      type: apiKey
      in: header
      name: X-Auth-Token
  parameters:
...
    ArtifactId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Unauthorized:
      description: Missing, invalid, or expired auth session
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Forbidden:
      description: Authenticated user does not have permission
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    NotFound:
      description: Record or file not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
    Conflict:
      description: Revision or state conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
...
  schemas:
    ApiErrorResponse:
      type: object
      required: [message, status, timestamp, issues, errorId]
      properties:
        message:
          type: string
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
        issues:
          type: array
          items: {}
        errorId:
          type: string
          description: Correlation id for support and log lookup
...
    RegisterRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        displayName:
          type: string
          maxLength: 80
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    UpdateMeRequest:
      type: object
      required: [displayName]
      properties:
        displayName:
          type: string
          maxLength: 80
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserDto"
    UserDto:
      type: object
      required: [id, email, displayName]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
...
    ArtifactRecord:
      type: object
      required: [id, projectId, name, modelJson, files, createdAt, updatedAt]
      properties:
        id:
          type: string
        projectId:
          type: string
        name:
          type: string
        modelJson:
          description: Source PSM model JSON.
        files:
          type: object
          additionalProperties:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SaveFileRequest:
      type: object
      required: [path]
      properties:
        path:
          type: string
        content:
          type: string
...